feat(animal, citizen): migrate enum value objects to enums
* replace value objects with enums for gender, species type, and weight unit in animal management
* add enums for civility, communication preference, and language preference in citizen management
* update related services and DTOs to use new enum types
* remove obsolete value object files
