+++ src/Beagl.Domain/AnimalManagement/Enums/Gender.cs
@@ -1,0 +1,24 @@
+// MIT License - Copyright (c) 2025 Jonathan St-Michel
+
+namespace Beagl.Domain.AnimalManagement.Enums;
+
+/// <summary>
+/// Represents the gender of an animal.
+/// </summary>
+public enum Gender
+{
+    /// <summary>
+    /// Unknown gender.
+    /// </summary>
+    Unknown = 0,
+
+    /// <summary>
+    /// Male gender.
+    /// </summary>
+    Male = 1,
+
+    /// <summary>
+    /// Female gender.
+    /// </summary>
+    Female = 2,
+}
