+++ src/Beagl.Domain/AnimalManagement/Enums/SpeciesType.cs
@@ -1,0 +1,24 @@
+// MIT License - Copyright (c) 2025 Jonathan St-Michel
+
+namespace Beagl.Domain.AnimalManagement.Enums;
+
+/// <summary>
+/// Enum for animal species type.
+/// </summary>
+public enum SpeciesType
+{
+    /// <summary>
+    /// Represents an unknown species type.
+    /// </summary>
+    Unknown = 0,
+
+    /// <summary>
+    /// Represents a cat.
+    /// </summary>
+    Cat = 1,
+
+    /// <summary>
+    /// Represents a dog.
+    /// </summary>
+    Dog = 2
+}
