+++ src/Beagl.Domain/AnimalManagement/Enums/WeightUnit.cs
@@ -1,0 +1,19 @@
+// MIT License - Copyright (c) 2025 Jonathan St-Michel
+
+namespace Beagl.Domain.AnimalManagement.Enums;
+
+/// <summary>
+/// Enum for weight measurement units.
+/// </summary>
+public enum WeightUnit
+{
+    /// <summary>
+    /// Represents weight in kilograms.
+    /// </summary>
+    Kilograms = 0,
+
+    /// <summary>
+    /// Represents weight in pounds.
+    /// </summary>
+    Pounds = 1
+}
