--- src/Beagl.Domain/AnimalManagement/ValueObjects/SpeciesType.cs
@@ -1,42 +1,0 @@
-// MIT License - Copyright (c) 2025 Jonathan St-Michel
-
-using System;
-
-namespace Beagl.Domain.AnimalManagement.ValueObjects;
-
-/// <summary>
-/// Enum for animal species type.
-/// </summary>
-public enum SpeciesType
-{
-    /// <summary>
-    /// Represents an unknown species type.
-    /// </summary>
-    Unknown = 0,
-
-    /// <summary>
-    /// Represents a cat.
-    /// </summary>
-    Cat = 1,
-
-    /// <summary>
-    /// Represents a dog.
-    /// </summary>
-    Dog = 2
-}
-
-/// <summary>
-/// Helper methods for the <see cref="SpeciesType"/> enum.
-/// </summary>
-public static class SpeciesTypeHelper
-{
-    /// <summary>
-    /// Converts an integer to a <see cref="SpeciesType"/> enum value.
-    /// </summary>
-    public static SpeciesType FromInt(int value)
-    {
-        if (!Enum.IsDefined(typeof(SpeciesType), value))
-            return SpeciesType.Unknown;
-        return (SpeciesType)value;
-    }
-}
