+++ src/Beagl.Domain/AnimalManagement/ValueObjects/Weight.cs
@@ -1,5 +1,7 @@
 // MIT License - Copyright (c) 2025 Jonathan St-Michel
 
+using Beagl.Domain.AnimalManagement.Enums;
+
 namespace Beagl.Domain.AnimalManagement.ValueObjects;
 
 /// <summary>
