+++ src/Beagl.WebApp/Services/StaticDataLocalizerService.cs
@@ -1,6 +1,6 @@
 // MIT License - Copyright (c) 2025 Jonathan St-Michel
 
-using Beagl.Domain.AnimalManagement.ValueObjects;
+using Beagl.Domain.AnimalManagement.Enums;
 using Microsoft.Extensions.Localization;
 
 namespace Beagl.WebApp.Services;
